third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/workers/baseurl/alpha/importScripts-in-sharedworker-expected.txt

This is a testharness.js-based test.
FAIL importScripts assert_equals: expected "gamma/script.js" but got "beta/script.js"
Harness: the test ran to completion.
